requirement.txt
langchain
langchain-openai
langchain-core
langchain_community
python-dotenv
langchain_huggingface
grandalf
PyPDF
langchain_experimental